requirements.txt
appdirs==1.4.3
attrs==16.3.0
Automat==0.5.0
boto3==1.4.4
botocore==1.5.26
constantly==15.1.0
docutils==0.13.1
futures==3.0.5
incremental==16.10.1
jmespath==0.9.2
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
s3transfer==0.1.10
six==1.10.0
Twisted==17.1.0
zope.interface==4.3.3
